Modified the Component names and provided comments at necessary places
