+++ src/styles/components/DoughnutChart.scss
@@ -1,4 +1,4 @@
-.circular-progress {
+.doughnut-chart {
     display: block;
     margin: 20px auto;
     max-width: 100%;
